+++ data-raw/rccsu2018.R
@@ -1,0 +1,2 @@
+library(nchsdata)
+usethis::use_data(rccsu2018, overwrite = TRUE)