+++ R/zzz.R
@@ -1,21 +1,21 @@
-.onAttach = function(libname, pkgname) {
-  txt = paste0(
-    "Before you can tabulate estimates, you have to specify which survey object"
-    , " you would like to analyze. You can do this in a couple of ways:"
-    , "\n\na) This package comes with a survey object for use in examples called"
-    , " 'namcs2019sv'. This object has selected variables from the NAMCS 2019 PUF survey."
-    , " To use this survey object:"
-    , "\n\nset_survey(namcs2019sv)"
-    , "\n\nb) If you have a survey object stored in a file:"
-    , "\n\nmysurvey = readRDS('file_name.rds')"
-    , "\n\nset_survey(mysurvey)"
-    , "\n\nFor info on how to create a survey object from a data frame, see"
-    , " ?survey::svydesign or ?survey::svrepdesign ."
-  )
-
-  txt = paste(strwrap(txt), collapse = "\n")
-  packageStartupMessage(txt)
-}
+# .onAttach = function(libname, pkgname) {
+#   txt = paste0(
+#     "Before you can tabulate estimates, you have to specify which survey object"
+#     , " you would like to analyze. You can do this in a couple of ways:"
+#     , "\n\na) This package comes with a survey object for use in examples called"
+#     , " 'namcs2019sv'. This object has selected variables from the NAMCS 2019 PUF survey."
+#     , " To use this survey object:"
+#     , "\n\nset_survey(namcs2019sv)"
+#     , "\n\nb) If you have a survey object stored in a file:"
+#     , "\n\nmysurvey = readRDS('file_name.rds')"
+#     , "\n\nset_survey(mysurvey)"
+#     , "\n\nFor info on how to create a survey object from a data frame, see"
+#     , " ?survey::svydesign or ?survey::svrepdesign ."
+#   )
+#
+#   txt = paste(strwrap(txt), collapse = "\n")
+#   packageStartupMessage(txt)
+# }
 
 env = new.env()