+++ NEWS.md
@@ -1,3 +1,5 @@
+# surveytable (development version)
+
 # surveytable 0.9.4
 
 * Optionally adjust p-values for multiple comparisons (`p_adjust` argument)