+++ NEWS.md
@@ -1,4 +1,4 @@
-# surveytable (development version)
+# surveytable 0.9.5
 
 * Another public use data file for use in examples: `rccsu2018`.
 * `set_opts()` replaces several other functions for setting options.